Revert to using int over short in Activity and derived classes

+++ Activities/GAScripted.cpp
@@ -200,7 +200,7 @@ int GAScripted::ReloadScripts()
 // Description:     Tells if a particular Scene supports this specific Activity on it.
 //                  Usually that means certain Area:s need to be defined in the Scene.
 
-bool GAScripted::SceneIsCompatible(Scene *pScene, short teams)
+bool GAScripted::SceneIsCompatible(Scene *pScene, int teams)
 {
     if (!GameActivity::SceneIsCompatible(pScene, teams))
         return false;
