+++ Activities/GameActivity.h
@@ -262,7 +262,7 @@ public:
 //                  be added to MovableMan already.
 // Return value:    Whether the focus switch was successful or not.
 
-	bool SwitchToActor(Actor *pActor, short player = 0, short team = 0) override;
+	bool SwitchToActor(Actor *pActor, int player = 0, int team = 0) override;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -274,7 +274,7 @@ public:
 //                  An actor pointer to skip in the sequence.
 // Return value:    None.
 
-	void SwitchToNextActor(short player, short team, Actor *pSkip = 0) override;
+	void SwitchToNextActor(int player, int team, Actor *pSkip = 0) override;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -286,7 +286,7 @@ public:
 //                  An actor pointer to skip in the sequence.
 // Return value:    None.
 
-	void SwitchToPrevActor(short player, short team, Actor *pSkip = 0) override;
+	void SwitchToPrevActor(int player, int team, Actor *pSkip = 0) override;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
